refactor(codemods): move existing utility functions to a shared place
These utility functions can be reused for the version 5 codemods, so I decided to move them to a shared place.
